% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptualSexStratifiedDistribution.R
\name{ConceptFigure}
\alias{ConceptFigure}
\title{Generating Conceptual Sex-Stratified Distributions as Shown in Xvarhet Manuscript}
\usage{
ConceptFigure(RAF, prp_m, length.OUT = 500, Yrange = 3, y1_vect = c(0, 1),
  y2_vect = c(-0.5, 0.5, 1, 1), y3_vect = c(0, 0, 0.8, 1.2),
  y4_vect = c(-0.5, 0.5, 0.8, 1.2), geno_sd = 0.2, cols = T)
}
\arguments{
\item{RAF}{a numeric between 0 and 1, the reference allele frequency of the
SNP, assumed to be equal in females and males.}

\item{prp_m}{a numeric between 0 and 1, giving the porportion of males
in the sample.}

\item{length.OUT}{an integer for the number of quantitative trait values
at which to evaluate; the default is set to 500.}

\item{Yrange}{a numeric for the maximum value of the quantitative trait to
be evaluated assuming the distribution is symmetric about zero;
the default is set to 3.}

\item{y1_vect}{a vector for the sex-stratified mean and standard deviation
in females and males under no sexual dimorphism; the default is c(0,1), with zero
mean and standard deviation one in both sexes.}

\item{y2_vect}{a vector for the sex-stratified mean and standard deviation
in females and males under sexual dimorphism in mean only; the default is
c(-0.5,0.5,1,1), with quantitative trait in female following a normal
distribution \eqn{N(\mu=-0.5, \sigma=1)}.}

\item{y3_vect}{a vector for the sex-stratified mean and standard deviation
in females and males under sexual dimorphism in variance only; the default
is c(0,0,0.8,1.2), with quantitative trait in female following a normal
distribution \eqn{N(\mu=0, \sigma=0.8)}.}

\item{y4_vect}{a vector for the sex-stratified mean and standard deviation
in females and males under sexual dimorphism in both mean and variance;
the default is c(-0.5,0.5,0.8,1.2), with quantitative trait in female following
a normal distribution \eqn{N(\mu=-0.5, \sigma=0.8)}.}

\item{geno_sd}{a numeric between 0 and the minimum of the sex-stratified
standard deviation, giving the contribution to standard deviation per
reference allele under no sexual dimorphism assuming an additive genetic
variance effect. The default is 0.2.}

\item{cols}{a logic indicating whether the outputted graphical object should be
in colour or gray style.}
}
\value{
a ggplot2 graph object
}
\description{
This function generates the conceptual distribution of a quantitative trait
stratified by sex and genotypes under the null hypothesis and alternative
hypothesis.
}
\examples{
#ConceptFigure(RAF=0.2, prp_m=0.1); # not run
#ConceptFigure(RAF=0.2, prp_m=0.1, cols = FALSE); # not run

}
\author{
Wei Q. Deng \email{deng@utstat.toronto.edu}
}
