% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeveneTest.R
\name{leveneTests}
\alias{leveneTests}
\title{Levene's test for variance homogeneity by SNP genotypes}
\usage{
leveneTests(GENO, SEX, PLINK = FALSE, Y, centre = "median", COV = NULL)
}
\arguments{
\item{GENO}{the genotype of a SNP, must be a vector of 0, 1, 2's indicating the number of reference alleles. The length of \code{GENO} should match that of \code{SEX}, \code{Y}, and any covariates.}

\item{SEX}{the genetic sex of individuals in the sample population, must be a vector of 1 and 2 or 0 and 1, depending on whether the PLINK sex code is used. Note that the default sex code is 1 for male and 2 for female in PLINK.}

\item{PLINK}{a logical indicating whether the SEX is coded following PLINK or not.}

\item{Y}{a vector of quantitative traits, such as human height.}

\item{centre}{a character indicating whether the absolute deviation should be calculated with respect to ``median'' or ``mean''.}

\item{COV}{a vector or matrix of covariates that are used to reduce bias due to confounding, such as age.}
}
\value{
a vector of Levene's test p-values according to levels specified by \code{GENO}, the
interaction between \code{GENO} and \code{SEX}, sex-specific Levene's test stratified by
\code{GENO}, and the Fisher's method to combine the sex-specific Levene's test \emph{p}-values.
}
\description{
The function takes as input the genotypes of a SNP (\code{GENO}), the sex (\code{SEX}), and a quantitative
   trait (\code{Y}) in a sample population, and possibly additional covariates. It should be noted
   that these variables must be of the same length. The function then
   returns the variance heterogeneity \emph{p}-values for the model \eqn{Y\sim G},
   \eqn{Y \sim G\times S}, the sex-specific results based on model \eqn{Y\sim G}, as well as that
   using Fisher's method to combine sex-specific results.
}
\note{
We recommend to quantile-normally transform \code{Y} to avoid ‘scale-effect’ where
the variance values tend to be proportional to mean values when stratified by \code{G}.
}
\examples{
N <- 5000
geno <- rbinom(N, 2, 0.3)
sex <- rbinom(N, 1, 0.5)
y <- rnorm(N)
cov <- matrix(rnorm(N*10), ncol=10)
leveneTests(GENO=geno, SEX=sex, Y=y, COV=cov)

}
\author{
Wei Q. Deng \email{deng@utstat.toronto.edu}
}
