% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationDesignI_T1E.R
\name{T1EsimDesign1}
\alias{T1EsimDesign1}
\title{Type I Error (T1E) Rates under Simulation Design I}
\usage{
T1EsimDesign1(Nf, Nm, MAF_F, MAF_M, betaE, betaS, betaSE, nbSim = 100,
  alpha = 0.05, nbDigits = 3, MisspecX = FALSE, test_type = "ALL")
}
\arguments{
\item{Nf}{an integer for the number of females in the sample population.}

\item{Nm}{an integer for the number of males in the sample population.}

\item{MAF_F}{a numerical between 0 and 0.5 for minor allele frequency of the simulated SNP in females.}

\item{MAF_M}{a numerical between 0 and 0.5 for minor allele frequency of the simulated SNP in males.}

\item{betaE}{a numerical value for the marginal environmental covariate effect on the simulated trait.}

\item{betaS}{a numerical value for the marginal sex effect on the simulated trait.}

\item{betaSE}{a numerical value for the interaction effect between sex and an environmental covariate.}

\item{nbSim}{an integer for the number of simulations.}

\item{alpha}{a numerical between 0 and 1 for the significance threshold, the default value is 0.05.}

\item{nbDigits}{an integer indicating the number of printed digits for empirical T1E rates.}

\item{MisspecX}{a logic indicating whether the T1E should be calculated under a
mis-specified X-inactivation status, i.e. model generated under X-inactivation
but tested under absence of X-inactivation.}

\item{test_type}{a character of either "ALL", printing all strategies, or "M1",
"M2", or "M3" for the 8 stratigies given the mean stage models, or "M3V3.2" printing
only the recommended test.}
}
\value{
a vector of type I error rates of the specified testing strategies.
}
\description{
This function returns the type I error rates of testing strategies
under the simulation conditions specified by a linear regression model.
The quantitative trait $Y$ is simulated from the following model without any genetic effect:
\eqn{Y =\beta_{E} E + \beta_S S + \beta_{SE} SE + \epsilon}, where \eqn{E} and \eqn{\epsilon} are assumed to follow a standard normal distribution.
}
\note{
We recommend to run the simulation in the background as it might take
long for more than 500 simulations on a 1.6 GHz Core.
}
\examples{
T1EsimDesign1(Nf=5000, Nm=5000, MAF_F=0.2, MAF_M=0.2, betaE=0.1,
betaS=0.2, betaSE=0, nbSim=5, alpha=0.05, nbDigits = 3, MisspecX =FALSE,
test_type = "ALL")

T1EsimDesign1(Nf=5000, Nm=5000, MAF_F=0.2, MAF_M=0.2, betaE=0.1,
betaS=0.2, betaSE=0, nbSim=50, alpha=0.05, nbDigits = 3, MisspecX =FALSE,
test_type = "M3V3.2")

T1EsimDesign1(Nf=5000, Nm=5000, MAF_F=0.2, MAF_M=0.2, betaE=0.1,
betaS=0.2, betaSE=0, nbSim=100, alpha=0.05, nbDigits = 3, Misspe
=TRUE,test_type = "ALL") # longer run time with 100 simulations

}
\author{
Wei Q. Deng \email{deng@utstat.toronto.edu}
}
